Update INSTALL-MAC.md
Incorrectly referenced Windows when meant Mac?

Hope I have interpreted this correctly? 18.06.1 is years old, is it really still a requirement? What if Balena Engine is installed on the Mac, will it use that instead and then Docker can be kept up to date? Running a version of Docker from 2018 is going to have a lot of impact on other work.