Added ability to SSH over tunnel.
To allow for a more easy device management experience for the openbalena users, I "added" the ability to use the tunnel for getting SSH acces. This can be enabled by simply adding the `--usetunnel` flag, which will tunnel to the device and then open an ssh session over it.

Change-type: minor