Enable dual stack support for balena tunnel
The `balena tunnel`commands uses `localhost` as the default listening
address. The problem is that on some dual stack networks (IPv4/IPv6),
this can result on the tunneling server only binding to `::1` and not
to `127.0.0.1`, potentially causing issues as this behavior is not
intuitive.

This PR replaces the default server name with `::` which setups the
server in dual-stack mode.

Change-type: patch